server/requirements.txt
fastapi
uvicorn
pydantic
langchain
psycopg2-binary
python-dotenv
pymupdf
